git commit -m "fix(security): 修复 JwtTokenProvider 启动失败问题"

+++ src/main/java/org/example/edumanagementservice/model/TeacherPermission.java
@@ -14,9 +14,9 @@ public class TeacherPermission {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    private Integer teacherId;
+    private Long teacherId;
 
     private Boolean canPublishCourse;
 }
